requirements.txt
paho-mqtt
graphitesend
pyserial
